feat: regen access token every hour & add git cliff
