+++ CHANGELOG.md
@@ -1,0 +1,24 @@
+# Changelog
+
+## [unreleased]
+
+### Bug Fixes
+
+-   Update all non-major dependencies
+
+## [1.0.5] - 2022-03-17
+
+### Bug Fixes
+
+-   Update all non-major dependencies
+-   Update all non-major dependencies
+
+### Features
+
+-   Add search manager
+
+## [1.0.0] - 2022-02-15
+
+### Features
+
+-   Add files
